+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+- Ctrl-C/KeyboardInterrupt results in exit code 130
 - Folders can be bundled
 - Test network availability even when not a scheduled run
 - Optionally per-server/per-folder --ctrl-args and --sync-args arguments